Parser v1.1
Added multytread

+++ DotaNet/Classes/Gamers/Gamer.cs
@@ -32,7 +32,7 @@ namespace DotaNet.Classes.Gamers
         {
             get
             {
-                return Wins / Looses;
+                return (double)Wins / (double)Looses;
             }
         }
         /// <summary>
